#12 moved backend and cli to their own directories
