+++ statistics-service/Dockerfile
