+++ statistics-service/api/event.go
@@ -2,8 +2,8 @@ package api
 
 import (
 	"github.com/gin-gonic/gin"
-	"github.com/keptn-sandbox/statistics-service/controller"
-	"github.com/keptn-sandbox/statistics-service/operations"
+	"github.com/keptn-sandbox/statistics-service/statistics-service/controller"
+	"github.com/keptn-sandbox/statistics-service/statistics-service/operations"
 	"net/http"
 )
