+++ statistics-service/controller/statistics_interface.go
@@ -1,8 +1,8 @@
 package controller
 
 import (
-	"github.com/keptn-sandbox/statistics-service/db"
-	"github.com/keptn-sandbox/statistics-service/operations"
+	"github.com/keptn-sandbox/statistics-service/statistics-service/db"
+	"github.com/keptn-sandbox/statistics-service/statistics-service/operations"
 	"time"
 )
