+++ statistics-service/db/mongodb_connection.go
