+++ statistics-service/db/statistics_mongodb_repo.go
@@ -3,7 +3,7 @@ package db
 import (
 	"context"
 	"github.com/globalsign/mgo/bson"
-	"github.com/keptn-sandbox/statistics-service/operations"
+	"github.com/keptn-sandbox/statistics-service/statistics-service/operations"
 	"go.mongodb.org/mongo-driver/mongo"
 	"time"
 )
