+++ statistics-service/docker/MANIFEST
