+++ statistics-service/docker/entrypoint.sh
