+++ statistics-service/docker/writeManifest.sh
