+++ statistics-service/docs/swagger.json
