+++ statistics-service/docs/swagger.yaml
