+++ statistics-service/main.go
