+++ statistics-service/operations/statistics_test.go
