+++ statistics-service/skaffold.yaml
